Split asset parser logic
