+++ src/extension.ts
@@ -1,12 +1,12 @@
 import * as vscode from 'vscode';
 import { initialize } from './initializer';
-import { UnityMetaDataParser } from './parser/unityMetaDataParser';
+import { UnityAssetConnector } from './parser/assetConnector';
 
 export function activate(context: vscode.ExtensionContext) {
 	initialize(context);
     
     // Test
-    UnityMetaDataParser.registerFileOpenHandler(context);
+    UnityAssetConnector.registerFileOpenHandler(context);
     
 	vscode.commands.executeCommand('workbench.view.extension.clover-activitybar');
 }
