+++ src/parser/assetParser.ts
@@ -1,5 +1,4 @@
 import fs = require('fs');
-import * as vscode from 'vscode';
 import * as Logger from '../vscodeUtils';
 
 interface MethodCall {
@@ -8,19 +7,10 @@ interface MethodCall {
     methodName: string;
 }
 
-interface MethodLocation {
+export interface MethodLocation {
     filePath: string;
     lineNumber: number;
     componentId: string;
-}
-
-interface UnityComponent {
-    type: string;
-    methodCalls?: MethodCall[];
-}
-
-interface UnityMetaData {
-    components: Map<string, UnityComponent>;
 }
 
 export class UnityMetaDataParser {
@@ -29,67 +19,10 @@ export class UnityMetaDataParser {
     private static readonly METHOD_CALLS_SECTION_PATTERN = /m_PersistentCalls:\s*\n\s*m_Calls:\s*((?:\s*-[^-]*)*)/g;
     private static readonly SINGLE_METHOD_CALL_PATTERN = /- m_Target: {fileID: (\d+)}\s*\n\s*m_TargetAssemblyTypeName: ([^,\n]+)(?:,[^\n]+)?\s*\n\s*m_MethodName: ([^\n]+)/g;
     private static readonly ASSEMBLY_TYPE_LINE_PATTERN = /\s*m_TargetAssemblyTypeName:/;
-    private static methodLocationCache: Map<string, MethodLocation[]> = new Map();
-
-    public static registerFileOpenHandler(context: vscode.ExtensionContext) {
-        let disposable = vscode.workspace.onDidOpenTextDocument(async (document) => {
-            if (this.isSupportedAsset(document.fileName)) {
-                const shouldParse = await vscode.window.showInformationMessage(
-                    'Unity asset file detected. Would you like to parse the metadata?',
-                    'Yes',
-                    'No'
-                );
-
-                if (shouldParse === 'Yes') {
-                    Logger.outputLog('Starting metadata parsing...');
-                    await this.parseMetaData(document.fileName);
-                    
-                    if (this.methodLocationCache.size > 0) {
-                        vscode.window.showInformationMessage(
-                            `Parsing complete! Found ${this.methodLocationCache.size} method calls.`
-                        );
-                    }
-                }
-            }
-        });
-
-        context.subscriptions.push(disposable);
-    }
 
     public static isSupportedAsset(filePath: string): boolean {
         const ext = filePath.toLowerCase().split('.').pop();
         return this.SUPPORTED_EXTENSIONS.includes(`.${ext}`);
-    }
-
-    public static validateMethod(namespaceName: string, className: string, methodName: string): { 
-        isValid: boolean; 
-        componentIds: string[];
-        foundIn: string[];
-    } {
-        const result = {
-            isValid: false,
-            componentIds: [] as string[],
-            foundIn: [] as string[]
-        };
-
-        this.methodLocationCache.forEach((locations, fullPath) => {
-            locations.forEach(loc => {
-                const [typeName, methodName] = fullPath.split('.');
-                if (typeName === namespaceName && 
-                    methodName === methodName) {
-                    result.isValid = true;
-                    result.componentIds.push(loc.componentId);
-                    result.foundIn.push(loc.filePath);
-                }
-            });
-        });
-
-        return result;
-    }
-
-    public static clearCache(): void {
-        this.methodLocationCache.clear();
-        Logger.outputLog('Method location cache has been cleared.');
     }
 
     private static getMethodFullPath(assemblyTypeName: string, methodName: string): string {
@@ -97,14 +30,9 @@ export class UnityMetaDataParser {
         return `${typeName}.${methodName}`;
     }
 
-    public static getMethodLocations(namespaceName: string, className: string, methodName: string): MethodLocation[] | undefined {
-        const fullPath = `${namespaceName}.${className}.${methodName}`;
-        return this.methodLocationCache.get(fullPath);
-    }
-
-    private static async extractMetaData(fileContent: string, filePath: string): Promise<void> {
+    private static async extractMetaData(fileContent: string, filePath: string): Promise<Map<string, MethodLocation[]>> {
+        const methodLocations = new Map<string, MethodLocation[]>();
         let match;
-        let currentLine = 1;
 
         while ((match = this.COMPONENT_PATTERN.exec(fileContent)) !== null) {
             const [fullMatch, id, type] = match;
@@ -146,33 +74,22 @@ export class UnityMetaDataParser {
                         componentId: id
                     };
 
-                    const existingLocations = this.methodLocationCache.get(fullPath) || [];
-                    this.methodLocationCache.set(fullPath, [...existingLocations, location]);
+                    const existingLocations = methodLocations.get(fullPath) || [];
+                    methodLocations.set(fullPath, [...existingLocations, location]);
                 }
             }
         }
 
-        Logger.outputLog('Cached method location information:');
-        if (this.methodLocationCache.size > 0) {
-            Logger.outputLog('----------------------------------------');
-            this.methodLocationCache.forEach((locations, fullPath) => {
-                Logger.outputLog(`\n[Method: ${fullPath}]`);
-                locations.forEach(loc => {
-                    Logger.outputLog(`  File: ${loc.filePath}`);
-                    Logger.outputLog(`  Line: ${loc.lineNumber}`);
-                    Logger.outputLog(`  Component ID: ${loc.componentId}`);
-                });
-            });
-            Logger.outputLog('----------------------------------------');
-        }
+        return methodLocations;
     }
 
-    public static async parseMetaData(filePath: string): Promise<void> {
+    public static async parseMetaData(filePath: string): Promise<Map<string, MethodLocation[]> | null> {
         try {
             const data = await fs.promises.readFile(filePath, { encoding: 'utf8' });
-            await this.extractMetaData(data, filePath);
+            return await this.extractMetaData(data, filePath);
         } catch (error) {
             Logger.outputLog(`Error parsing Unity asset at ${filePath}: ${error}`);
+            return null;
         }
     }
-}
+}
